segunda habilidad y sprites en direccion correcta

+++ Proyecto_1/src/Components/CircleColliderComponent.hpp
@@ -2,11 +2,11 @@
 #define CIRCLECOLLIDERCOMPONENT_HPP
 
 struct CircleColliderComponent{
- int radius;
- int width;
- int height;
+ double radius;
+ double width;
+ double height;
 
- CircleColliderComponent(int radius = 0,int width = 0,int height = 0){
+ CircleColliderComponent(double radius = 0,double width = 0,double height = 0){
     this->radius = radius;
     this->width = width;
     this->height = height;
